Fix module docs link in lib with no explicit intf
Voodoo needs to be consistent about the URL where it serves the module
docs of a library and the URL where it requests it. For the serve URL,
it always uses the module structure as given by dune. However, for
the request URL, it used to use an unwrapped version of the module
structure whenever the library doesn't have an explicit library
interface. That's why this commit removes that unwrapping of the module
structure.

Co-authored-by: Jon Ludlam <jon@recoil.org>